+++ src/main/java/org/jenkinsci/plugins/fstrigger/service/FSTriggerFileNameCheckedModifiedService.java
@@ -1,7 +1,7 @@
 package org.jenkinsci.plugins.fstrigger.service;
 
-import org.jenkinsci.lib.xtrigger.XTriggerException;
-import org.jenkinsci.lib.xtrigger.XTriggerLog;
+import org.jenkinsci.plugins.xtriggerapi.XTriggerException;
+import org.jenkinsci.plugins.xtriggerapi.XTriggerLog;
 import org.jenkinsci.plugins.fstrigger.core.FSTriggerContentFileType;
 import org.jenkinsci.plugins.fstrigger.triggers.FileNameTriggerInfo;